Initial commit. Write README enumerating project goals.
kpmalloc is less about the malloc and more about the implications
of nesting allocators, with privileged variants capable of attaching
to hardware PMP to replace a traditional MMU / virtual memory
subsystem in embedded systems. Write a README which describes
this further.

Meanwhile, copy the Makefile from a different project and start
tweaking it to our needs.

Signed-off-by: Jay Lang <jaytlang@mit.edu>