Add datastructure definitions, functions, and the global header
True to K&R, set up an aligned struct that acts as part of a
circularly linked list. Elegant and effective, and also adaptable
for a number of different approaches from here. Additionally,
spec out the actual calls to malloc and free themselves, in direct
accordance to their typical Linux-y implementation.

Set up (platform-specific) initialization specs, as well as the
debug printing routine once we tweak the makefile to enable this.
Hopefully this is straightforward...

Signed-off-by: Jay Lang <jaytlang@mit.edu>