Spec out test suite functionality in a new fns
Write a series of helper functions which allows the manipulation
of a blackbox structure -- allowing us to keep track of what's
going on inside of kpmalloc without actually tweaking that code
too much. This is the varlist, and will be interacted with via
the promised REPL.

Place specs for this promised REPL into fns.h and verify everything
is syntactically correct. Implementation inbound.

Signed-off-by: Jay Lang <jaytlang@mit.edu>