Finish REPL, institute bug fixes to get this thing working
Get the REPL to work! Some malformed loops were present
throughout the code; fix these and ensure the allocator
exhibits proper merging/slicing behavior. So far it
appears this is intact.

Valgrind is still kinda mad at me, but I'll bet money it's
an issue with the prompt rather than the actual kpmalloc
subsystem. I'm going to try and write a second test application
with this thing on a separate branch, and merge it in if
successful. Hopefully this will diagnose the issue in full.

Signed-off-by: Jay Lang <jaytlang@mit.edu>