Create a scratchpad for messing around with test applications
So I've killed off the prompt in this branch in favor with a
simple main.c -- test applications can be easily built and
linked to kpmalloc here. I'm pretty sure that everything works,
judging from how happy valgrind is, but also valgrind is poorly
equipped to instrument the mmap/munmap + internal management
that kpmalloc does. This should at least be helpful to
customize applications and further test out the functionality
of kpmalloc.

Signed-off-by: Jay Lang <jaytlang@mit.edu>